refactor : [코드리뷰 반영] 불필요한 static 제거

+++ src/main/java/lotto/domain/Prize.java
@@ -37,7 +37,7 @@ public enum Prize {
                 .collect(Collectors.toList());
     }
 
-    public int calculatePriceMoney(int count) {
+    public long calculatePriceMoney(long count) {
         return this.prizeMoney * count;
     }
